+++ components/Map.tsx
@@ -1,12 +1,18 @@
 import React, { useEffect, useState } from "react";
-import { Text, View } from "react-native";
+import { ActivityIndicator, Text, View } from "react-native";
 import MapView, { Marker, PROVIDER_DEFAULT } from "react-native-maps";
-import { calculateRegion, generateMarkersFromData } from "@/app/lib/map";
+import {
+  calculateDriverTimes,
+  calculateRegion,
+  generateMarkersFromData,
+} from "@/app/lib/map";
 import { useDriverStore, useLocationStore } from "@/store";
 import { Driver, MarkerData } from "@/types/type";
 import { icons } from "@/constants";
+import { useFetch } from "@/app/lib/fetch";
+import MapViewDirections from "react-native-maps-directions";
 
-const drivers = [
+/* const drivers = [
   {
     id: 1,
     first_name: "James",
@@ -51,9 +57,10 @@ const drivers = [
     car_seats: 4,
     rating: 4.9,
   },
-];
+]; */
 
 const Map = () => {
+  const { data: drivers, loading, error } = useFetch<Driver[]>("/(api)/driver");
   const { selectedDriver, setDrivers } = useDriverStore();
   const [markers, setMarkers] = useState<MarkerData[]>([]);
 
@@ -66,18 +73,46 @@ const Map = () => {
 
   useEffect(() => {
     // TODO remove this later
-    setDrivers(drivers as MarkerData[]);
+    // setDrivers(drivers as MarkerData[]);
     if (Array.isArray(drivers)) {
       if (!userLatitude || !userLongitude) return;
-    }
 
-    const newMarkers = generateMarkersFromData({
-      data: drivers,
-      userLatitude,
-      userLongitude,
-    });
-    setMarkers(newMarkers);
+      const newMarkers = generateMarkersFromData({
+        data: drivers!,
+        userLatitude,
+        userLongitude,
+      });
+      setMarkers(newMarkers);
+    }
   }, [drivers]);
+
+  useEffect(() => {
+    if (markers.length > 0 && destinationLatitude && destinationLongitude) {
+      calculateDriverTimes({
+        markers,
+        userLongitude,
+        userLatitude,
+        destinationLatitude,
+        destinationLongitude,
+      }).then((drivers) => {
+        setDrivers(drivers as MarkerData[]);
+      });
+    }
+  }, [markers, destinationLatitude, destinationLatitude]);
+
+  if (loading || !userLatitude || !userLongitude)
+    return (
+      <View className="flex justify-between items-center w-full">
+        <ActivityIndicator size="small" color="#000" />
+      </View>
+    );
+
+  if (error)
+    return (
+      <View className="flex justify-between items-center w-full">
+        <Text>Error: {error}</Text>
+      </View>
+    );
 
   const region = calculateRegion({
     userLatitude,
@@ -94,7 +129,7 @@ const Map = () => {
       mapType="mutedStandard"
       showsPointsOfInterest={false}
       initialRegion={region}
-      showsUserLocation={true}
+      // showsUserLocation={true}
       userInterfaceStyle="light"
     >
       {markers.map((marker) => (
@@ -110,6 +145,33 @@ const Map = () => {
           }
         ></Marker>
       ))}
+      {destinationLatitude && destinationLongitude && (
+        <>
+          <Marker
+            key="destination"
+            coordinate={{
+              latitude: destinationLatitude,
+              longitude: destinationLongitude,
+            }}
+            title="Destination"
+            image={icons.pin}
+          />
+
+          <MapViewDirections
+            origin={{
+              latitude: userLatitude,
+              longitude: userLongitude,
+            }}
+            destination={{
+              latitude: destinationLatitude,
+              longitude: destinationLongitude,
+            }}
+            apikey={process.env.EXPO_PUBLIC_GOOGLE_API_KEY!}
+            strokeColor="#0286ff"
+            strokeWidth={2}
+          />
+        </>
+      )}
     </MapView>
   );
 };