+++ components/Payment.tsx
@@ -1,11 +1,32 @@
-import { Alert } from "react-native";
+import { Alert, Image, View } from "react-native";
 import { PaymentSheetError, useStripe } from "@stripe/stripe-react-native";
 import CustomButton from "./CustomButton";
 import { useState } from "react";
+import { fetchAPI } from "@/app/lib/fetch";
+import { PaymentProps } from "@/types/type";
+import { useLocationStore } from "@/store";
+import { useAuth } from "@clerk/clerk-expo";
+import ReactNativeModal from "react-native-modal";
+import { images } from "@/constants";
 
-const Payment = () => {
+const Payment = ({
+  fullName,
+  email,
+  amount,
+  driverId,
+  rideTime,
+}: PaymentProps) => {
   const { initPaymentSheet, presentPaymentSheet } = useStripe();
-  const [success, setSuccess] = useState<boolean>(false);
+  const { userId } = useAuth();
+  const [success, setSuccess] = useState<boolean>(true);
+  const {
+    userAddress,
+    userLongitude,
+    userLatitude,
+    destinationLatitude,
+    destinationLongitude,
+    destinationAddress,
+  } = useLocationStore();
 
   const initializePaymentSheet = async () => {
     const { error } = await initPaymentSheet({
@@ -15,12 +36,76 @@ const Payment = () => {
           amount: 1099,
           currencyCode: "USD",
         },
-        // confirmHandler: confirmHandler,
-        confirmHandler: () => {},
+        confirmHandler: async (paymentMethod, _, intentCreationCallback) => {
+          const { paymentIntent, customer } = await fetchAPI(
+            "/(api)/(stripe)/create",
+            {
+              method: "POST",
+              headers: {
+                "Content-Type": "application/json",
+              },
+              body: JSON.stringify({
+                name: fullName || email.split("@")[0],
+                email,
+                amount,
+                paymentMethodId: paymentMethod.id,
+              }),
+            }
+          );
+
+          if (paymentIntent.client_secret) {
+            const { result } = await fetchAPI("/(api)/(stripe)/pay", {
+              method: "POST",
+              headers: {
+                "Content-Type": "application/json",
+              },
+              body: JSON.stringify({
+                payment_method_id: paymentMethod.id,
+                payment_intent_id: paymentIntent.id,
+                customer_id: customer,
+              }),
+            });
+            if (result.client_secret) {
+              await fetchAPI("/(api)/ride/create", {
+                method: "POST",
+                headers: {
+                  "Content-Type": "application/json",
+                },
+                body: JSON.stringify({
+                  origin_address: userAddress,
+                  destination_address: destinationAddress,
+                  origin_latitude: userLatitude,
+                  origin_longitude: userLongitude,
+                  destination_latitude: destinationLatitude,
+                  destination_longitude: destinationLongitude,
+                  ride_time: rideTime.toFixed(0),
+                  fare_price: parseInt(amount) * 100,
+                  payment_status: "paid",
+                  driver_id: driverId,
+                  user_id: userId,
+                }),
+              });
+
+              intentCreationCallback({
+                clientSecret: result.client_secret,
+              });
+            }
+          }
+
+          // Call the `intentCreationCallback` with your server response's client secret or error
+          /* const { clientSecret, error } = await response.json();
+    if (clientSecret) {
+      intentCreationCallback({ clientSecret });
+    } else {
+      intentCreationCallback({ error });
+    } */
+        },
       },
+      returnURL: "myapp://book-ride",
     });
     if (error) {
       // handle error
+      console.log(error);
     }
   };
 
@@ -50,6 +135,17 @@ const Payment = () => {
         className="my-10"
         onPress={openPaymentSheet}
       />
+      <ReactNativeModal
+        isVisible={success}
+        onBackdropPress={() => setSuccess(false)}
+      >
+        <View
+          className="flex flex-col items-center justify-center
+        bg-white p-7 rounded-2xl"
+        >
+          <Image source={images.check} className="w-28 h-28 mt-5" />
+        </View>
+      </ReactNativeModal>
     </>
   );
 };