+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - stripe intent completed
 - stripe intent completed
+- stripe intent completed
 - stripe component completed
 - stripe integration in progress
 - book ride payment component  implementation